+++ include/nstd/core/str.h
@@ -3,10 +3,16 @@
 #include "../nstd.h"
 #include "cstr.h"
 #include "def.h"
+#include "optional.h"
 #include "range.h"
 #include "slice.h"
 
 /// An immutable unowned view into a UTF-8 encoded byte string.
+///
+/// # Safety
+///
+/// The user of this structure must ensure that the pointed-to data remains valid UTF-8, and
+/// unmodified while an instance of this structure is in use.
 typedef struct {
     /// A raw pointer to the string's data.
     const NSTDByte *ptr;
@@ -229,7 +235,7 @@ NSTDAPI NSTDUInt nstd_core_str_byte_len(const NSTDStr *str);
 /// # Safety
 ///
 /// This operation can cause undefined behavior in the event that `str`'s data is invalid.
-NSTDAPI NSTDUnichar nstd_core_str_get_char(const NSTDStr *str, NSTDUInt pos);
+NSTDAPI NSTDUnichar nstd_core_str_get(const NSTDStr *str, NSTDUInt pos);
 
 /// Creates a substring of an existing string slice.
 ///
@@ -268,11 +274,9 @@ NSTDAPI NSTDStr nstd_core_str_substr(const NSTDStr *str, NSTDURange range);
 ///
 /// - `const NSTDStr *str` - The string slice.
 ///
-/// - `NSTDErrorCode *errc` - Set to nonzero on error.
-///
 /// # Returns
 ///
-/// `NSTDFloat32 v` - The parsed 32-bit floating-point value.
+/// `NSTDOptionalFloat32 v` - The parsed value, or none on error.
 ///
 /// # Panics
 ///
@@ -282,18 +286,16 @@ NSTDAPI NSTDStr nstd_core_str_substr(const NSTDStr *str, NSTDURange range);
 /// # Safety
 ///
 /// This operation can cause undefined behavior in the event that `str`'s data is invalid.
-NSTDAPI NSTDFloat32 nstd_core_str_to_f32(const NSTDStr *str, NSTDErrorCode *errc);
+NSTDAPI NSTDOptionalFloat32 nstd_core_str_to_f32(const NSTDStr *str);
 /// Attempts to parse a string slice as an `NSTDFloat64`.
 ///
 /// # Parameters:
 ///
 /// - `const NSTDStr *str` - The string slice.
 ///
-/// - `NSTDErrorCode *errc` - Set to nonzero on error.
-///
 /// # Returns
 ///
-/// `NSTDFloat64 v` - The parsed 64-bit floating-point value.
+/// `NSTDOptionalFloat64 v` - The parsed value, or none on error.
 ///
 /// # Panics
 ///
@@ -303,18 +305,16 @@ NSTDAPI NSTDFloat32 nstd_core_str_to_f32(const NSTDStr *str, NSTDErrorCode *errc
 /// # Safety
 ///
 /// This operation can cause undefined behavior in the event that `str`'s data is invalid.
-NSTDAPI NSTDFloat64 nstd_core_str_to_f64(const NSTDStr *str, NSTDErrorCode *errc);
+NSTDAPI NSTDOptionalFloat64 nstd_core_str_to_f64(const NSTDStr *str);
 /// Attempts to parse a string slice as an `NSTDInt`.
 ///
 /// # Parameters:
 ///
 /// - `const NSTDStr *str` - The string slice.
 ///
-/// - `NSTDErrorCode *errc` - Set to nonzero on error.
-///
 /// # Returns
 ///
-/// `NSTDInt v` - The parsed arch-bit signed integral value.
+/// `NSTDOptionalInt v` - The parsed value, or none on error.
 ///
 /// # Panics
 ///
@@ -324,18 +324,16 @@ NSTDAPI NSTDFloat64 nstd_core_str_to_f64(const NSTDStr *str, NSTDErrorCode *errc
 /// # Safety
 ///
 /// This operation can cause undefined behavior in the event that `str`'s data is invalid.
-NSTDAPI NSTDInt nstd_core_str_to_int(const NSTDStr *str, NSTDErrorCode *errc);
+NSTDAPI NSTDOptionalInt nstd_core_str_to_int(const NSTDStr *str);
 /// Attempts to parse a string slice as an `NSTDUInt`.
 ///
 /// # Parameters:
 ///
 /// - `const NSTDStr *str` - The string slice.
 ///
-/// - `NSTDErrorCode *errc` - Set to nonzero on error.
-///
 /// # Returns
 ///
-/// `NSTDUInt v` - The parsed arch-bit unsigned integral value.
+/// `NSTDOptionalUInt v` - The parsed value, or none on error.
 ///
 /// # Panics
 ///
@@ -345,18 +343,16 @@ NSTDAPI NSTDInt nstd_core_str_to_int(const NSTDStr *str, NSTDErrorCode *errc);
 /// # Safety
 ///
 /// This operation can cause undefined behavior in the event that `str`'s data is invalid.
-NSTDAPI NSTDUInt nstd_core_str_to_uint(const NSTDStr *str, NSTDErrorCode *errc);
+NSTDAPI NSTDOptionalUInt nstd_core_str_to_uint(const NSTDStr *str);
 /// Attempts to parse a string slice as an `NSTDInt8`.
 ///
 /// # Parameters:
 ///
 /// - `const NSTDStr *str` - The string slice.
 ///
-/// - `NSTDErrorCode *errc` - Set to nonzero on error.
-///
 /// # Returns
 ///
-/// `NSTDInt8 v` - The parsed 8-bit signed integral value.
+/// `NSTDOptionalInt8 v` - The parsed value, or none on error.
 ///
 /// # Panics
 ///
@@ -366,18 +362,16 @@ NSTDAPI NSTDUInt nstd_core_str_to_uint(const NSTDStr *str, NSTDErrorCode *errc);
 /// # Safety
 ///
 /// This operation can cause undefined behavior in the event that `str`'s data is invalid.
-NSTDAPI NSTDInt8 nstd_core_str_to_i8(const NSTDStr *str, NSTDErrorCode *errc);
+NSTDAPI NSTDOptionalInt8 nstd_core_str_to_i8(const NSTDStr *str);
 /// Attempts to parse a string slice as an `NSTDUInt8`.
 ///
 /// # Parameters:
 ///
 /// - `const NSTDStr *str` - The string slice.
 ///
-/// - `NSTDErrorCode *errc` - Set to nonzero on error.
-///
 /// # Returns
 ///
-/// `NSTDUInt8 v` - The parsed 8-bit unsigned integral value.
+/// `NSTDOptionalUInt8 v` - The parsed value, or none on error.
 ///
 /// # Panics
 ///
@@ -387,18 +381,16 @@ NSTDAPI NSTDInt8 nstd_core_str_to_i8(const NSTDStr *str, NSTDErrorCode *errc);
 /// # Safety
 ///
 /// This operation can cause undefined behavior in the event that `str`'s data is invalid.
-NSTDAPI NSTDUInt8 nstd_core_str_to_u8(const NSTDStr *str, NSTDErrorCode *errc);
+NSTDAPI NSTDOptionalUInt8 nstd_core_str_to_u8(const NSTDStr *str);
 /// Attempts to parse a string slice as an `NSTDInt16`.
 ///
 /// # Parameters:
 ///
 /// - `const NSTDStr *str` - The string slice.
 ///
-/// - `NSTDErrorCode *errc` - Set to nonzero on error.
-///
 /// # Returns
 ///
-/// `NSTDInt16 v` - The parsed 16-bit signed integral value.
+/// `NSTDOptionalInt16 v` - The parsed value, or none on error.
 ///
 /// # Panics
 ///
@@ -408,18 +400,16 @@ NSTDAPI NSTDUInt8 nstd_core_str_to_u8(const NSTDStr *str, NSTDErrorCode *errc);
 /// # Safety
 ///
 /// This operation can cause undefined behavior in the event that `str`'s data is invalid.
-NSTDAPI NSTDInt16 nstd_core_str_to_i16(const NSTDStr *str, NSTDErrorCode *errc);
+NSTDAPI NSTDOptionalInt16 nstd_core_str_to_i16(const NSTDStr *str);
 /// Attempts to parse a string slice as an `NSTDUInt16`.
 ///
 /// # Parameters:
 ///
 /// - `const NSTDStr *str` - The string slice.
 ///
-/// - `NSTDErrorCode *errc` - Set to nonzero on error.
-///
 /// # Returns
 ///
-/// `NSTDUInt16 v` - The parsed 16-bit unsigned integral value.
+/// `NSTDOptionalUInt16 v` - The parsed value, or none on error.
 ///
 /// # Panics
 ///
@@ -429,18 +419,16 @@ NSTDAPI NSTDInt16 nstd_core_str_to_i16(const NSTDStr *str, NSTDErrorCode *errc);
 /// # Safety
 ///
 /// This operation can cause undefined behavior in the event that `str`'s data is invalid.
-NSTDAPI NSTDUInt16 nstd_core_str_to_u16(const NSTDStr *str, NSTDErrorCode *errc);
+NSTDAPI NSTDOptionalUInt16 nstd_core_str_to_u16(const NSTDStr *str);
 /// Attempts to parse a string slice as an `NSTDInt32`.
 ///
 /// # Parameters:
 ///
 /// - `const NSTDStr *str` - The string slice.
 ///
-/// - `NSTDErrorCode *errc` - Set to nonzero on error.
-///
 /// # Returns
 ///
-/// `NSTDInt32 v` - The parsed 32-bit signed integral value.
+/// `NSTDOptionalInt32 v` - The parsed value, or none on error.
 ///
 /// # Panics
 ///
@@ -450,18 +438,16 @@ NSTDAPI NSTDUInt16 nstd_core_str_to_u16(const NSTDStr *str, NSTDErrorCode *errc)
 /// # Safety
 ///
 /// This operation can cause undefined behavior in the event that `str`'s data is invalid.
-NSTDAPI NSTDInt32 nstd_core_str_to_i32(const NSTDStr *str, NSTDErrorCode *errc);
+NSTDAPI NSTDOptionalInt32 nstd_core_str_to_i32(const NSTDStr *str);
 /// Attempts to parse a string slice as an `NSTDUInt32`.
 ///
 /// # Parameters:
 ///
 /// - `const NSTDStr *str` - The string slice.
 ///
-/// - `NSTDErrorCode *errc` - Set to nonzero on error.
-///
 /// # Returns
 ///
-/// `NSTDUInt32 v` - The parsed 32-bit unsigned integral value.
+/// `NSTDOptionalUInt32 v` - The parsed value, or none on error.
 ///
 /// # Panics
 ///
@@ -471,18 +457,16 @@ NSTDAPI NSTDInt32 nstd_core_str_to_i32(const NSTDStr *str, NSTDErrorCode *errc);
 /// # Safety
 ///
 /// This operation can cause undefined behavior in the event that `str`'s data is invalid.
-NSTDAPI NSTDUInt32 nstd_core_str_to_u32(const NSTDStr *str, NSTDErrorCode *errc);
+NSTDAPI NSTDOptionalUInt32 nstd_core_str_to_u32(const NSTDStr *str);
 /// Attempts to parse a string slice as an `NSTDInt64`.
 ///
 /// # Parameters:
 ///
 /// - `const NSTDStr *str` - The string slice.
 ///
-/// - `NSTDErrorCode *errc` - Set to nonzero on error.
-///
 /// # Returns
 ///
-/// `NSTDInt64 v` - The parsed 64-bit signed integral value.
+/// `NSTDOptionalInt64 v` - The parsed value, or none on error.
 ///
 /// # Panics
 ///
@@ -492,18 +476,16 @@ NSTDAPI NSTDUInt32 nstd_core_str_to_u32(const NSTDStr *str, NSTDErrorCode *errc)
 /// # Safety
 ///
 /// This operation can cause undefined behavior in the event that `str`'s data is invalid.
-NSTDAPI NSTDInt64 nstd_core_str_to_i64(const NSTDStr *str, NSTDErrorCode *errc);
+NSTDAPI NSTDOptionalInt64 nstd_core_str_to_i64(const NSTDStr *str);
 /// Attempts to parse a string slice as an `NSTDUInt64`.
 ///
 /// # Parameters:
 ///
 /// - `const NSTDStr *str` - The string slice.
 ///
-/// - `NSTDErrorCode *errc` - Set to nonzero on error.
-///
 /// # Returns
 ///
-/// `NSTDUInt64 v` - The parsed 64-bit unsigned integral value.
+/// `NSTDOptionalUInt64 v` - The parsed value, or none on error.
 ///
 /// # Panics
 ///
@@ -513,9 +495,15 @@ NSTDAPI NSTDInt64 nstd_core_str_to_i64(const NSTDStr *str, NSTDErrorCode *errc);
 /// # Safety
 ///
 /// This operation can cause undefined behavior in the event that `str`'s data is invalid.
-NSTDAPI NSTDUInt64 nstd_core_str_to_u64(const NSTDStr *str, NSTDErrorCode *errc);
+NSTDAPI NSTDOptionalUInt64 nstd_core_str_to_u64(const NSTDStr *str);
 
 /// An unowned view into a UTF-8 encoded byte string.
+///
+/// # Safety
+///
+/// The user of this structure must ensure that the pointed-to data remains valid UTF-8, unmodified,
+/// and unreferenced in any other code while an instance of this structure is in use, else data
+/// races may occur.
 typedef struct {
     /// A raw pointer to the string's data.
     NSTDByte *ptr;
@@ -749,7 +737,7 @@ NSTDAPI NSTDUInt nstd_core_str_mut_byte_len(const NSTDStrMut *str);
 /// # Safety
 ///
 /// This operation can cause undefined behavior in the event that `str`'s data is invalid.
-NSTDAPI NSTDUnichar nstd_core_str_mut_get_char(const NSTDStrMut *str, NSTDUInt pos);
+NSTDAPI NSTDUnichar nstd_core_str_mut_get(const NSTDStrMut *str, NSTDUInt pos);
 
 /// Creates a substring of an existing string slice.
 ///
@@ -788,11 +776,9 @@ NSTDAPI NSTDStrMut nstd_core_str_mut_substr(NSTDStrMut *str, NSTDURange range);
 ///
 /// - `const NSTDStrMut *str` - The string slice.
 ///
-/// - `NSTDErrorCode *errc` - Set to nonzero on error.
-///
 /// # Returns
 ///
-/// `NSTDFloat32 v` - The parsed 32-bit floating-point value.
+/// `NSTDOptionalFloat32 v` - The parsed value, or none on error.
 ///
 /// # Panics
 ///
@@ -802,18 +788,16 @@ NSTDAPI NSTDStrMut nstd_core_str_mut_substr(NSTDStrMut *str, NSTDURange range);
 /// # Safety
 ///
 /// This operation can cause undefined behavior in the event that `str`'s data is invalid.
-NSTDAPI NSTDFloat32 nstd_core_str_mut_to_f32(const NSTDStrMut *str, NSTDErrorCode *errc);
+NSTDAPI NSTDOptionalFloat32 nstd_core_str_mut_to_f32(const NSTDStrMut *str);
 /// Attempts to parse a string slice as an `NSTDFloat64`.
 ///
 /// # Parameters:
 ///
 /// - `const NSTDStrMut *str` - The string slice.
 ///
-/// - `NSTDErrorCode *errc` - Set to nonzero on error.
-///
 /// # Returns
 ///
-/// `NSTDFloat64 v` - The parsed 64-bit floating-point value.
+/// `NSTDOptionalFloat64 v` - The parsed value, or none on error.
 ///
 /// # Panics
 ///
@@ -823,18 +807,16 @@ NSTDAPI NSTDFloat32 nstd_core_str_mut_to_f32(const NSTDStrMut *str, NSTDErrorCod
 /// # Safety
 ///
 /// This operation can cause undefined behavior in the event that `str`'s data is invalid.
-NSTDAPI NSTDFloat64 nstd_core_str_mut_to_f64(const NSTDStrMut *str, NSTDErrorCode *errc);
+NSTDAPI NSTDOptionalFloat64 nstd_core_str_mut_to_f64(const NSTDStrMut *str);
 /// Attempts to parse a string slice as an `NSTDInt`.
 ///
 /// # Parameters:
 ///
 /// - `const NSTDStrMut *str` - The string slice.
 ///
-/// - `NSTDErrorCode *errc` - Set to nonzero on error.
-///
 /// # Returns
 ///
-/// `NSTDInt v` - The parsed arch-bit signed integral value.
+/// `NSTDOptionalInt v` - The parsed value, or none on error.
 ///
 /// # Panics
 ///
@@ -844,18 +826,16 @@ NSTDAPI NSTDFloat64 nstd_core_str_mut_to_f64(const NSTDStrMut *str, NSTDErrorCod
 /// # Safety
 ///
 /// This operation can cause undefined behavior in the event that `str`'s data is invalid.
-NSTDAPI NSTDInt nstd_core_str_mut_to_int(const NSTDStrMut *str, NSTDErrorCode *errc);
+NSTDAPI NSTDOptionalInt nstd_core_str_mut_to_int(const NSTDStrMut *str);
 /// Attempts to parse a string slice as an `NSTDUInt`.
 ///
 /// # Parameters:
 ///
 /// - `const NSTDStrMut *str` - The string slice.
 ///
-/// - `NSTDErrorCode *errc` - Set to nonzero on error.
-///
 /// # Returns
 ///
-/// `NSTDUInt v` - The parsed arch-bit unsigned integral value.
+/// `NSTDOptionalUInt v` - The parsed value, or none on error.
 ///
 /// # Panics
 ///
@@ -865,18 +845,16 @@ NSTDAPI NSTDInt nstd_core_str_mut_to_int(const NSTDStrMut *str, NSTDErrorCode *e
 /// # Safety
 ///
 /// This operation can cause undefined behavior in the event that `str`'s data is invalid.
-NSTDAPI NSTDUInt nstd_core_str_mut_to_uint(const NSTDStrMut *str, NSTDErrorCode *errc);
+NSTDAPI NSTDOptionalUInt nstd_core_str_mut_to_uint(const NSTDStrMut *str);
 /// Attempts to parse a string slice as an `NSTDInt8`.
 ///
 /// # Parameters:
 ///
 /// - `const NSTDStrMut *str` - The string slice.
 ///
-/// - `NSTDErrorCode *errc` - Set to nonzero on error.
-///
 /// # Returns
 ///
-/// `NSTDInt8 v` - The parsed 8-bit signed integral value.
+/// `NSTDOptionalInt8 v` - The parsed value, or none on error.
 ///
 /// # Panics
 ///
@@ -886,18 +864,16 @@ NSTDAPI NSTDUInt nstd_core_str_mut_to_uint(const NSTDStrMut *str, NSTDErrorCode 
 /// # Safety
 ///
 /// This operation can cause undefined behavior in the event that `str`'s data is invalid.
-NSTDAPI NSTDInt8 nstd_core_str_mut_to_i8(const NSTDStrMut *str, NSTDErrorCode *errc);
+NSTDAPI NSTDOptionalInt8 nstd_core_str_mut_to_i8(const NSTDStrMut *str);
 /// Attempts to parse a string slice as an `NSTDUInt8`.
 ///
 /// # Parameters:
 ///
 /// - `const NSTDStrMut *str` - The string slice.
 ///
-/// - `NSTDErrorCode *errc` - Set to nonzero on error.
-///
 /// # Returns
 ///
-/// `NSTDUInt8 v` - The parsed 8-bit unsigned integral value.
+/// `NSTDOptionalUInt8 v` - The parsed value, or none on error.
 ///
 /// # Panics
 ///
@@ -907,18 +883,16 @@ NSTDAPI NSTDInt8 nstd_core_str_mut_to_i8(const NSTDStrMut *str, NSTDErrorCode *e
 /// # Safety
 ///
 /// This operation can cause undefined behavior in the event that `str`'s data is invalid.
-NSTDAPI NSTDUInt8 nstd_core_str_mut_to_u8(const NSTDStrMut *str, NSTDErrorCode *errc);
+NSTDAPI NSTDOptionalUInt8 nstd_core_str_mut_to_u8(const NSTDStrMut *str);
 /// Attempts to parse a string slice as an `NSTDInt16`.
 ///
 /// # Parameters:
 ///
 /// - `const NSTDStrMut *str` - The string slice.
 ///
-/// - `NSTDErrorCode *errc` - Set to nonzero on error.
-///
 /// # Returns
 ///
-/// `NSTDInt16 v` - The parsed 16-bit signed integral value.
+/// `NSTDOptionalInt16 v` - The parsed value, or none on error.
 ///
 /// # Panics
 ///
@@ -928,18 +902,16 @@ NSTDAPI NSTDUInt8 nstd_core_str_mut_to_u8(const NSTDStrMut *str, NSTDErrorCode *
 /// # Safety
 ///
 /// This operation can cause undefined behavior in the event that `str`'s data is invalid.
-NSTDAPI NSTDInt16 nstd_core_str_mut_to_i16(const NSTDStrMut *str, NSTDErrorCode *errc);
+NSTDAPI NSTDOptionalInt16 nstd_core_str_mut_to_i16(const NSTDStrMut *str);
 /// Attempts to parse a string slice as an `NSTDUInt16`.
 ///
 /// # Parameters:
 ///
 /// - `const NSTDStrMut *str` - The string slice.
 ///
-/// - `NSTDErrorCode *errc` - Set to nonzero on error.
-///
 /// # Returns
 ///
-/// `NSTDUInt16 v` - The parsed 16-bit unsigned integral value.
+/// `NSTDOptionalUInt16 v` - The parsed value, or none on error.
 ///
 /// # Panics
 ///
@@ -949,18 +921,16 @@ NSTDAPI NSTDInt16 nstd_core_str_mut_to_i16(const NSTDStrMut *str, NSTDErrorCode 
 /// # Safety
 ///
 /// This operation can cause undefined behavior in the event that `str`'s data is invalid.
-NSTDAPI NSTDUInt16 nstd_core_str_mut_to_u16(const NSTDStrMut *str, NSTDErrorCode *errc);
+NSTDAPI NSTDOptionalUInt16 nstd_core_str_mut_to_u16(const NSTDStrMut *str);
 /// Attempts to parse a string slice as an `NSTDInt32`.
 ///
 /// # Parameters:
 ///
 /// - `const NSTDStrMut *str` - The string slice.
 ///
-/// - `NSTDErrorCode *errc` - Set to nonzero on error.
-///
 /// # Returns
 ///
-/// `NSTDInt32 v` - The parsed 32-bit signed integral value.
+/// `NSTDOptionalInt32 v` - The parsed value, or none on error.
 ///
 /// # Panics
 ///
@@ -970,18 +940,16 @@ NSTDAPI NSTDUInt16 nstd_core_str_mut_to_u16(const NSTDStrMut *str, NSTDErrorCode
 /// # Safety
 ///
 /// This operation can cause undefined behavior in the event that `str`'s data is invalid.
-NSTDAPI NSTDInt32 nstd_core_str_mut_to_i32(const NSTDStrMut *str, NSTDErrorCode *errc);
+NSTDAPI NSTDOptionalInt32 nstd_core_str_mut_to_i32(const NSTDStrMut *str);
 /// Attempts to parse a string slice as an `NSTDUInt32`.
 ///
 /// # Parameters:
 ///
 /// - `const NSTDStrMut *str` - The string slice.
 ///
-/// - `NSTDErrorCode *errc` - Set to nonzero on error.
-///
 /// # Returns
 ///
-/// `NSTDUInt32 v` - The parsed 32-bit unsigned integral value.
+/// `NSTDOptionalUInt32 v` - The parsed value, or none on error.
 ///
 /// # Panics
 ///
@@ -991,18 +959,16 @@ NSTDAPI NSTDInt32 nstd_core_str_mut_to_i32(const NSTDStrMut *str, NSTDErrorCode 
 /// # Safety
 ///
 /// This operation can cause undefined behavior in the event that `str`'s data is invalid.
-NSTDAPI NSTDUInt32 nstd_core_str_mut_to_u32(const NSTDStrMut *str, NSTDErrorCode *errc);
+NSTDAPI NSTDOptionalUInt32 nstd_core_str_mut_to_u32(const NSTDStrMut *str);
 /// Attempts to parse a string slice as an `NSTDInt64`.
 ///
 /// # Parameters:
 ///
 /// - `const NSTDStrMut *str` - The string slice.
 ///
-/// - `NSTDErrorCode *errc` - Set to nonzero on error.
-///
 /// # Returns
 ///
-/// `NSTDInt64 v` - The parsed 64-bit signed integral value.
+/// `NSTDOptionalInt64 v` - The parsed value, or none on error.
 ///
 /// # Panics
 ///
@@ -1012,18 +978,16 @@ NSTDAPI NSTDUInt32 nstd_core_str_mut_to_u32(const NSTDStrMut *str, NSTDErrorCode
 /// # Safety
 ///
 /// This operation can cause undefined behavior in the event that `str`'s data is invalid.
-NSTDAPI NSTDInt64 nstd_core_str_mut_to_i64(const NSTDStrMut *str, NSTDErrorCode *errc);
+NSTDAPI NSTDOptionalInt64 nstd_core_str_mut_to_i64(const NSTDStrMut *str);
 /// Attempts to parse a string slice as an `NSTDUInt64`.
 ///
 /// # Parameters:
 ///
 /// - `const NSTDStrMut *str` - The string slice.
 ///
-/// - `NSTDErrorCode *errc` - Set to nonzero on error.
-///
 /// # Returns
 ///
-/// `NSTDUInt64 v` - The parsed 64-bit unsigned integral value.
+/// `NSTDOptionalUInt64 v` - The parsed value, or none on error.
 ///
 /// # Panics
 ///
@@ -1033,6 +997,6 @@ NSTDAPI NSTDInt64 nstd_core_str_mut_to_i64(const NSTDStrMut *str, NSTDErrorCode 
 /// # Safety
 ///
 /// This operation can cause undefined behavior in the event that `str`'s data is invalid.
-NSTDAPI NSTDUInt64 nstd_core_str_mut_to_u64(const NSTDStrMut *str, NSTDErrorCode *errc);
+NSTDAPI NSTDOptionalUInt64 nstd_core_str_mut_to_u64(const NSTDStrMut *str);
 
 #endif