+++ include/nstd/heap_ptr.h
@@ -102,7 +102,7 @@ NSTDAPI NSTDAnyMut nstd_heap_ptr_get_mut(NSTDHeapPtr *hptr);
 ///
 /// # Panics
 ///
-/// This operation may panic if getting a handle to the heap fails.
+/// Panics if freeing the heap memory fails.
 NSTDAPI void nstd_heap_ptr_free(NSTDHeapPtr hptr);
 
 #endif