+++ include/nstd/os/windows.h
@@ -1,4 +1,5 @@
 #ifndef NSTD_OS_WINDOWS_H
 #define NSTD_OS_WINDOWS_H
 #include "windows/alloc.h"
+#include "windows/shared_lib.h"
 #endif